fix: refactor library structure and fix tests
