+++ src/core/common/error.c
