+++ src/core/common/identity.c
