+++ src/core/memory/manifest.c
