+++ src/core/memory/memory_flash.c
