+++ src/core/memory/memory_impl.c
