+++ src/core/memory/memory_objects.c
