+++ src/core/memory/simple_manifest.c
