+++ src/core/network/connection_config.c
