+++ src/core/network/subscriber.c
