+++ src/core/network/writer.c
