+++ src/core/network/writer_buffered.c
