+++ src/core/security/cryptoauthlib.c
