+++ src/core/security/tinycrypt.c
