+++ src/core/security/verifier.c
