+++ src/network/coap/libcoap/async_libcoap.h
@@ -1,6 +1,8 @@
 #ifndef ASYNC_LIBCOAP_H_
 #define ASYNC_LIBCOAP_H_
 
+#ifdef WITH_CONNECTION_LIBCOAP
+
 #include <libpull/common.h>
 #include <libpull/network.h>
 
@@ -14,4 +16,5 @@ void loop(txp_ctx* ctx, uint32_t timeout);
 
 void break_loop(txp_ctx* ctx);
 
-#endif /* ASYNC_LIBCOAP_H_ */
+#endif /* WITH_CONNECTION_LIBCOAP */
+#endif /* ASYNC_LIBCOAP_H_ */
