+++ test/Makefile.am
@@ -1,7 +1,7 @@
 AUTOMAKE_OPTIONS = subdir-objects
 AM_CFLAGS = -Wall -I. -I../include -I../ext/tinycrypt/lib/include -I../ext/tinydtls -DHAVE_CONFIG_H
 AM_CFLAGS += -I../build/posix -Isupport -I../ext/unity/src -I../ext/libcoap/include -Imocks
-AM_CFLAGS += -DHAVE_CONFIG_H -DUSE_TESTING_KEYS=1
+AM_CFLAGS += -DHAVE_CONFIG_H -DUSE_TESTING_KEYS=1 -DWITH_CONNECTION_LIBCOAP=1 -DWITH_UPDATE_AGENT=1
 
 if COND_GCOV
   MAYBE_COVERAGE= -fprofile-arcs -ftest-coverage
@@ -16,27 +16,27 @@ LIBS = ../ext/libcoap/.libs/libcoap-1.a ../ext/tinydtls/libtinydtls.a
 LIBS += ../ext/tinycrypt/lib/libtinycrypt.a
 
 UNITY_SRC = ../ext/unity/src/unity.c
-COMMON_SRC = ../src/common/error.c ../src/common/identity.c
-MEMORY_SRC = ../src/memory/memory_objects.c 
-SECURITY_SRC = ../src/security/verifier.c
-NETWORK_SRC = ../src/network/receiver.c ../src/network/subscriber.c ../src/network/connection_config.c ../src/network/writer.c
+COMMON_SRC = ../src/core/common/error.c ../src/core/common/identity.c
+MEMORY_SRC = ../src/core/memory/memory_objects.c 
+SECURITY_SRC = ../src/core/security/verifier.c
+NETWORK_SRC = ../src/core/network/receiver.c ../src/core/network/subscriber.c ../src/core/network/connection_config.c ../src/core/network/writer.c
 SUPPORT_SRC = support/external_variables.c support/utils.c
 
-AGENTS_SRC = ../agents/update.c ../agents/bootloader_agent.c
+AGENTS_SRC = ../src/agents/update.c ../src/agents/bootloader_agent.c
 
-NETWORK_IMPL = ../build/posix/connection_libcoap.c ../build/posix/async_libcoap.c
-MEMORY_IMPL = ../src/memory/memory_impl.c ../build/posix/memory_posix.c
-MANIFEST_IMPL = ../src/memory/manifest.c ../src/memory/simple_manifest.c
+NETWORK_IMPL = ../src/network/coap/libcoap/connection_libcoap.c ../src/network/coap/libcoap/async_libcoap.c
+MEMORY_IMPL = ../src/core/memory/memory_impl.c ../build/posix/memory_posix.c
+MANIFEST_IMPL = ../src/core/memory/manifest.c ../src/core/memory/simple_manifest.c
 
 ALL_SRC = $(UNITY_SRC) $(COMMON_SRC) $(MEMORY_SRC) $(SECURITY_SRC) $(MANIFEST_IMPL) $(NETWORK_SRC)
 ALL_SRC += $(NETWORK_IMPL) $(MEMORY_IMPL) $(SUPPORT_SRC)
 
 MOCK_MEMORY = mocks/memory_mock.c ../build/posix/memory_posix.c
-MOCK_MANIFEST = mocks/manifest_mock.c ../src/memory/simple_manifest.c
+MOCK_MANIFEST = mocks/manifest_mock.c ../src/core/memory/simple_manifest.c
 NO_MOCK_SRC = $(UNITY_SRC) $(COMMON_SRC) $(SECURITY_SRC) $(NETWORK_SRC) $(NETWORK_IMPL) $(MEMORY_SRC) $(SUPPORT_SRC)
 
-TINYCRYPT_SRC = ../src/security/tinycrypt.c ../ext/tinycrypt/lib/source/ecc_platform_specific.c
-TINYDTLS_SRC = ../src/security/tinydtls.c
+TINYCRYPT_SRC = ../src/core/security/tinycrypt.c ../ext/tinycrypt/lib/source/ecc_platform_specific.c
+TINYDTLS_SRC = ../src/core/security/tinydtls.c
 
 memory_tests = test_memory test_memory_objects test_manifest
 agents_tests = test_update test_bootloader_agent
